moved search_tree object to a separated file. added a depth-first clean method.
