+++ cpp/src/branch_and_bound/mip_node.hpp
@@ -41,43 +41,8 @@ inline bool inactive_status(node_status_t status)
 template <typename i_t, typename f_t>
 class mip_node_t {
  public:
-  ~mip_node_t()
-  {
-    // Iterative teardown to avoid stack overflow on deep trees.
-    // Detach all descendants breadth-first, then destroy them as leaves.
-    // vector::push_back can throw bad_alloc; the catch-all keeps the destructor
-    // exception-free. Under OOM, any not-yet-detached descendants are destroyed
-    // via the recursive unique_ptr chain in `children` as this frame unwinds.
-    try {
-      std::vector<std::unique_ptr<mip_node_t>> nodes;
-      for (auto& c : children) {
-        if (c) { nodes.push_back(std::move(c)); }
-      }
-      // nodes.size() grows so that this loop only terminates when only leaves remain
-      for (size_t i = 0; i < nodes.size(); ++i) {
-        for (auto& c : nodes[i]->children) {
-          if (c) { nodes.push_back(std::move(c)); }
-        }
-      }
-
-      // scope-exit ensure destruction of all detached leaves
-    } catch (const std::exception& e) {
-      // fprintf to stderr is allocation-free and cannot throw; using the
-      // project logger here would risk a secondary bad_alloc that would
-      // escape the destructor and re-introduce std::terminate.
-      std::fprintf(stderr,
-                   "mip_node_t destructor: iterative teardown failed (%s); falling back to "
-                   "recursive unique_ptr destruction.\n",
-                   e.what());
-    } catch (...) {
-      std::fprintf(stderr,
-                   "mip_node_t destructor: iterative teardown failed (unknown exception); "
-                   "falling back to recursive unique_ptr destruction.\n");
-    }
-  }
-
-  mip_node_t(mip_node_t&&) noexcept            = default;
-  mip_node_t& operator=(mip_node_t&&) noexcept = default;
+  mip_node_t(mip_node_t&&)            = default;
+  mip_node_t& operator=(mip_node_t&&) = default;
 
   mip_node_t()
     : status(node_status_t::PENDING),
@@ -363,100 +328,5 @@ void remove_fathomed_nodes(std::vector<mip_node_t<i_t, f_t>*>& stack)
     }
   }
 }
-
-template <typename i_t, typename f_t>
-class search_tree_t {
- public:
-  search_tree_t() : num_nodes(0) {}
-
-  search_tree_t(mip_node_t<i_t, f_t>&& node) : root(std::move(node)), num_nodes(0) {}
-
-  void update(mip_node_t<i_t, f_t>* node_ptr, node_status_t status)
-  {
-    std::lock_guard<omp_mutex_t> lock(mutex);
-    std::vector<mip_node_t<i_t, f_t>*> stack;
-    node_ptr->set_status(status, stack);
-    remove_fathomed_nodes(stack);
-  }
-
-  void branch(mip_node_t<i_t, f_t>* parent_node,
-              const i_t branch_var,
-              const f_t fractional_val,
-              const i_t integer_infeasible,
-              const std::vector<variable_status_t>& parent_vstatus,
-              const lp_problem_t<i_t, f_t>& original_lp,
-              logger_t& log)
-  {
-    i_t id = num_nodes.fetch_add(2);
-
-    auto down_child = std::make_unique<mip_node_t<i_t, f_t>>(original_lp,
-                                                             parent_node,
-                                                             ++id,
-                                                             branch_var,
-                                                             branch_direction_t::DOWN,
-                                                             fractional_val,
-                                                             integer_infeasible,
-                                                             parent_vstatus);
-    graphviz_edge(log,
-                  parent_node,
-                  down_child.get(),
-                  branch_var,
-                  branch_direction_t::DOWN,
-                  std::floor(fractional_val));
-
-    auto up_child = std::make_unique<mip_node_t<i_t, f_t>>(original_lp,
-                                                           parent_node,
-                                                           ++id,
-                                                           branch_var,
-                                                           branch_direction_t::UP,
-                                                           fractional_val,
-                                                           integer_infeasible,
-                                                           parent_vstatus);
-
-    graphviz_edge(log,
-                  parent_node,
-                  up_child.get(),
-                  branch_var,
-                  branch_direction_t::UP,
-                  std::ceil(fractional_val));
-
-    assert(parent_vstatus.size() == original_lp.num_cols);
-    parent_node->add_children(std::move(down_child),
-                              std::move(up_child));  // child pointers moved into the tree
-  }
-
-  void graphviz_node(logger_t& log,
-                     const mip_node_t<i_t, f_t>* node_ptr,
-                     const std::string label,
-                     const f_t val)
-  {
-    if (write_graphviz) {
-      log.printf("Node%d [label=\"%s %.16e\"]\n", node_ptr->node_id, label.c_str(), val);
-    }
-  }
-
-  void graphviz_edge(logger_t& log,
-                     const mip_node_t<i_t, f_t>* origin_ptr,
-                     const mip_node_t<i_t, f_t>* dest_ptr,
-                     const i_t branch_var,
-                     branch_direction_t branch_dir,
-                     const f_t bound)
-  {
-    if (write_graphviz) {
-      log.printf("Node%d -> Node%d [label=\"x%d %s %e\"]\n",
-                 origin_ptr->node_id,
-                 dest_ptr->node_id,
-                 branch_var,
-                 branch_dir == branch_direction_t::DOWN ? "<=" : ">=",
-                 bound);
-    }
-  }
-
-  mip_node_t<i_t, f_t> root;
-  omp_mutex_t mutex;
-  omp_atomic_t<i_t> num_nodes;
-
-  static constexpr bool write_graphviz = false;
-};
 
 }  // namespace cuopt::linear_programming::dual_simplex
